Move groups code to own file
